feat: 하드웨어 연결 api연동
